there is an x on the screen in the trade state that i can click and it exits the trade

+++ aspirant_application/Game.Shoppes.cpp
@@ -54,7 +54,7 @@ namespace game::shoppe
 		}
 	}
 
-	const Descriptor& GetDescriptor(int index)
+	const Descriptor& GetDescriptor(size_t index)
 	{
 		Initialize();
 		return descriptors[index];
